controllers/krate/metadata: Move `get_version_readme()` to dedicated module
