+++ src/controllers/version.rs
@@ -1,5 +1,6 @@
 pub mod downloads;
 pub mod metadata;
+pub mod readme;
 pub mod yank;
 
 use axum::extract::{FromRequestParts, Path};
