+++ src/router.rs
@@ -37,7 +37,7 @@ pub fn build_axum_router(state: AppState) -> Router<()> {
             version::metadata::find_version,
             version::metadata::update_version
         ))
-        .routes(routes!(krate::metadata::get_version_readme))
+        .routes(routes!(version::readme::get_version_readme))
         .routes(routes!(version::metadata::get_version_dependencies))
         .routes(routes!(version::downloads::get_version_downloads))
         .routes(routes!(version::metadata::get_version_authors))
